requirements.txt
google-api-python-client==2.1.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
PyYAML==5.4.1